moved the haversine distance function to the utils submodule
